+++ src/quacc/settings.py
@@ -49,7 +49,7 @@ class QuaccSettings(BaseSettings):
     # ---------------------------
     # Data Store Settings
     # ---------------------------
-    RESULTS_STORE: str = Field(
+    PRIMARY_STORE: str = Field(
         None,
         description="String-based JSON representation of the primary Maggma data store where calculation results will be stored. Taken from the `.to_json()` method of the corresponding Store object.",
     )