+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-- Added a `RESULTS_STORE` quacc setting that, when specified, will automatically store all results in the specified Maggma store.
+- Added a `PRIMARY_STORE` quacc setting that, when specified, will automatically store all results in the specified Maggma store.
 
 ### Changed